# src/BindingChaos.Web/api-spec.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "BindingChaos.Web.Gateway",
    "description": "BindingChaos Web Gateway API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "HTTP dev (default)"
    },
    {
      "url": "https://localhost:4001",
      "description": "HTTPS dev"
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Creates a new project.",
        "operationId": "createProject",
        "requestBody": {
          "x-name": "request",
          "description": "The project creation request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "The created project identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfString"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Retrieves a paginated list of projects for the specified user group.",
        "operationId": "getProjectsForUserGroup",
        "parameters": [
          {
            "name": "userGroupId",
            "in": "query",
            "description": "The user group identifier to filter by.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "Page.Number",
            "in": "query",
            "description": "The current page number (1-based).",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "Page.Size",
            "in": "query",
            "description": "The requested page size.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "sort",
            "in": "query",
            "style": "form",
            "explode": true,
            "description": "Parsed sort descriptors bound from the querystring parameter named 'sort'.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SortDescriptor"
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated response containing project list items.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPaginatedResponseOfProjectListItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Retrieves a single project by identifier.",
        "operationId": "getProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The project details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfProjectResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/amendments": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Proposes a new amendment for a project.",
        "operationId": "proposeProjectAmendment",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "201": {
            "description": "The created amendment identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/amendments/{amendmentId}/contests": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Contests an active amendment on a project.",
        "operationId": "contestProjectAmendment",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "amendmentId",
            "in": "path",
            "required": true,
            "description": "The amendment identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": "204 No Content on success."
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/inquiries": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Raises a new inquiry against a project.",
        "operationId": "raiseProjectInquiry",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The raise inquiry request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RaiseProjectInquiryRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "The created inquiry identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfString"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Lists inquiries for a project.",
        "operationId": "getProjectInquiries",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "Page.Number",
            "in": "query",
            "description": "The current page number (1-based).",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "Page.Size",
            "in": "query",
            "description": "The requested page size.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "sort",
            "in": "query",
            "style": "form",
            "explode": true,
            "description": "Parsed sort descriptors bound from the querystring parameter named 'sort'.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SortDescriptor"
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "A paginated list of inquiries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfPaginatedResponseOfProjectInquiryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/inquiries/{inquiryId}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Gets a single project inquiry.",
        "operationId": "getProjectInquiry",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "inquiryId",
            "in": "path",
            "required": true,
            "description": "The inquiry identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "The inquiry details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfProjectInquiryResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Updates the body of an inquiry, resetting it to open status.",
        "operationId": "updateProjectInquiry",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "inquiryId",
            "in": "path",
            "required": true,
            "description": "The inquiry identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The update request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectInquiryRequest"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "204": {
            "description": "No content on success."
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/inquiries/{inquiryId}/responses": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Submits a user group response to an inquiry.",
        "operationId": "respondToProjectInquiry",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "inquiryId",
            "in": "path",
            "required": true,
            "description": "The inquiry identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The response request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondToProjectInquiryRequest"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "204": {
            "description": "No content on success."
          }
        }
      }
    },
    "/api/v1/projects/{projectId}/inquiries/{inquiryId}/resolutions": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Resolves an inquiry, accepting the user group's response.",
        "operationId": "resolveProjectInquiry",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "description": "The project identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "inquiryId",
            "in": "path",
            "required": true,
            "description": "The inquiry identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": "No content on success."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ApiResponseOfString": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "description": "Standardized API response envelope as defined in ADR-00016.",
            "additionalProperties": false,
            "properties": {
              "data": {
                "type": "string",
                "description": "Gets or sets the response data.",
                "nullable": true
              }
            }
          }
        ]
      },
      "ApiResponse": {
        "type": "object",
        "description": "Static factory methods for creating API responses.",
        "additionalProperties": false
      },
...
      "CreateProjectRequest": {
        "type": "object",
        "description": "Request model for creating a project in a user group.",
        "additionalProperties": false,
        "properties": {
          "userGroupId": {
            "type": "string",
            "description": "The owning user group identifier."
          },
          "title": {
            "type": "string",
            "description": "The project title."
          },
          "description": {
            "type": "string",
            "description": "The project description."
          }
        }
      },
...
    }
  }
}